# api.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: YardRaffles
servers:
  - url: 'http://localhost:8888'
paths:
...
  /api/users/get/{id}:
    get:
      tags:
        - users
      summary: "Получить информацию о пользователе"
      operationId: getUserById
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        schema:
            type: string
      - in: path
        name: id
        schema:
            type: integer
        required: true
        description: id пользователя
      responses:
        200:
          description: "Объект с записью"
          content:
            application/json:
              schema:
                $ref: '#/definitions/UserData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
        404:
          description: "Пользователь не найден в БД"
...
  /api/users/block/{id}:
    put:
      tags:
      - users
      summary: Заблокировать пользователя
      operationId: blockUser
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) для расчета стоимости индивидуального посещения. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        type: string
      - in: query
        name: reason
        type: string
        required: true
        description: "Goofy ski ski"
      - in: path
        name: id
        type: integer
        required: true
        description: Id пользователя
        
      responses:
        200:
          description: "Операция прошла успешно"
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  /api/users/unblock/{id}:
    put:
      tags:
      - users
      summary: Разблокировать пользователя
      operationId: unblockUser
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) для расчета стоимости индивидуального посещения. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        type: string
      - in: path
        name: id
        type: integer
        required: true
        description: Id пользователя
        
      responses:
        200:
          description: "Операция прошла успешно"
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  /api/users/accept/{id}:
    put:
      tags:
      - users
      summary: Подтверждает регистрацию пользователя
      operationId: acceptUser
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) для расчета стоимости индивидуального посещения. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        type: string
      - in: path
        name: id
        type: integer
        required: true
        description: Id пользователя
        
      responses:
        200:
          description: "Операция прошла успешно"
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  
  # USERS END
  
  # RAFFLES START
  /api/raffles/get/{id}:
    get:
      tags:
      - raffles
      summary: Получить информацию о раффле
      operationId: getRaffleData
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        type: string
      - in: path
        name: id
        type: integer
        required: true
        description: id раффла
        
      responses:
        200:
          description: "Объект с информацией о раффле"
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/RaffleData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
        404:
          description: "Раффл не найден"
  /api/raffles/create:
    post:
      tags:
      - raffles
      summary: Создать раффл
      operationId: createRaffle
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) для расчета стоимости индивидуального посещения. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        type: string
      - in: query
        name: title
        type: string
        required: true
        description: "Название"
      - in: query
        name: work_name
        type: string
        required: true
        description: "Рабочее название"
      - in: query
        name: images
        type: string
        required: false
        description: "Hash изображения"
      - in: query
        name: message
        type: string
        required: true
        description: "Сообщение в боте"
      - in: query
        name: link
        type: string
        required: true
        description: "Ссылка на раффл"
      - in: query
        name: sizes
        type: string
        required: false
        description: "Список размеров"
      - in: query
        name: publication_date
        type: string
        required: false
        description: "Время публикации (если не указано, то сразу)"
      - in: query
        name: close_date
        type: string
        required: false
        description: "Время закртия регистрации на раффл (если не указано, то закрывать вручную)"
      - in: query
        name: results_date
        type: string
        required: false
        description: "Время сообщения результатов"
      - in: query
        name: profit
        type: number
        required: false
        description: "Возможная прибыль"
      - in: body
        name: tags
        required: false
        description: "Массив объектов id тэгов"
        schema:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
      
      responses:
        200:
          description: "Объект с созданной записью"
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/RaffleData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  /api/raffles/update/{id}:
    put:
      tags:
      - raffles
      summary: Обновить раффл
      operationId: updateRaffle
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) для расчета стоимости индивидуального посещения. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        type: string
      - in: path
        name: id
        type: integer
        required: true
        description: Id раффла
      - in: query
        name: title
        type: string
        required: true
        description: "Название"
      - in: query
        name: work_name
        type: string
        required: true
        description: "Рабочее название"
      - in: query
        name: images
        type: string
        required: false
        description: "Hash изображения"
      - in: query
        name: message
        type: string
        required: true
        description: "Сообщение в боте"
      - in: query
        name: link
        type: string
        required: true
        description: "Ссылка на раффл"
      - in: query
        name: sizes
        type: string
        required: false
        description: "Список размеров"
      - in: query
        name: publication_date
        type: string
        required: false
        description: "Время публикации (если не указано, то сразу)"
      - in: query
        name: close_date
        type: string
        required: false
        description: "Время закртия регистрации на раффл (если не указано, то закрывать вручную)"
      - in: query
        name: results_date
        type: string
        required: false
        description: "Время сообщения результатов"
      - in: query
        name: profit
        type: number
        required: false
        description: "Возможная прибыль"
      - in: body
        name: tags
        required: false
        description: "Массив объектов id тэгов"
        schema:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
      responses:
        200:
          description: "Объект с созданной записью"
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/RaffleData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  /api/raffles/delete/{id}:
    delete:
      tags:
      - raffles
      summary: Удалить запись раффла
      operationId: deleteRaffle
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) для расчета стоимости индивидуального посещения. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        type: string
      - in: path
        name: id
        type: integer
        required: true
        description: Id записи
        
      responses:
        200:
          description: "Операция прошла успешно"
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  /api/raffles/getRaffles/{page}:
    get:
      tags:
      - raffles
      summary: "Список раффлов"
      operationId: getRaffleList
      produces:
      - application/json
      parameters:
      - in: path
        name: page
        type: integer
        required: true
        description: Страница
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) для расчета стоимости индивидуального посещения. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        type: string
      - in: query
        name: search
        description: "Поисковое слово"
        required: false
        type: string
      - in: query
        name: date_from
        description: "ГГГГ-ММ-ДД"
        required: false
        type: string
      - in: query
        name: date_to
        description: "ГГГГ-ММ-ДД"
        required: false
        type: string
      - in: query
        name: status
        required: false
        description: "Массив статусов"
        type: array
        items:
          type: integer
      - in: query
        name: users
        required: false
        description: "Массив пользователей"
        type: array
        items:
          type: integer
      - in: query
        name: sort
        type: string
        required: false
        description: "Параметр сортировки. Столбец, по которому нужна сортировка. Например, date_s."
      - in: query
        name: sort_dir
        type: string
        required: false
        description: "Тип сортировки: asc или desc"
      responses:
        200:
          description: "Массив билетов. Если записей нет, то выдаем пустой массив"
          schema:
            type: object
            properties:
              count:
                type: integer
                example: 12
              rows:
                type: array
                items:
                  $ref: '#/definitions/CommonRaffleData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  # RAFFLES END
  
  # ANNOUCE START
...
  /api/announces/create:
    post:
      tags:
      - announcements
      summary: Создать объялвение
      operationId: createAnnounce
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) для расчета стоимости индивидуального посещения. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        type: string
      - in: query
        name: title
        type: string
        required: true
        description: "Название"
      - in: query
        name: work_name
        type: string
        required: true
        description: "Рабочее название"
      - in: query
        name: images
        type: string
        required: false
        description: "Hash изображения"
      - in: query
        name: message
        type: string
        required: true
        description: "Сообщение в боте"
      - in: query
        name: publication_date
        type: string
        required: false
        description: "Время публикации (если не указано, то сразу)"
      - in: body
        name: tags
        required: false
        description: "Массив объектов id тэгов"
        schema:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
      responses:
        200:
          description: "Объект с созданной записью"
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/AnnounceData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
...